Assistant checkpoint: Update PostCSS config to use CommonJS
Assistant generated file changes:
- postcss.config.cjs: Convert PostCSS config to CommonJS

---

User prompt:

Running the code in my project results in the following:

```
kage.json' contains "type": "module". To treat it as a CommonJS script, rename it to use the '.cjs' file extension.
Import trace for requested module:
./src/styles/tailwind.css.webpack[javascript/auto]!=!./node_modules/next/dist/build/webpack/loaders/css-loader/src/index.js??ruleSet[1].rules[14].oneOf[12].use[2]!./node_modules/next/dist/build/webpack/loaders/postcss-loader/src/index.js??ruleSet[1].rules[14].oneOf[12].use[3]!./src/styles/tailwind.css
./src/styles/tailwind.css
 ⨯ ./src/styles/tailwind.css.webpack[javascript/auto]!=!./node_modules/next/dist/build/webpack/loaders/css-loader/src/index.js??ruleSet[1].rules[14].oneOf[12].use[2]!./node_modules/next/dist/build/webpack/loaders/postcss-loader/src/index.js??ruleSet[1].rules[14].oneOf[12].use[3]!./src/styles/tailwind.css
ReferenceError: module is not defined in ES module scope
This file is being treated as an ES module because it has a '.js' file extension and '/home/runner/Personal-Site-Test/package.json' contains "type": "module". To treat it as a CommonJS script, rename it to use the '.cjs' file extension.
Import trace for requested module:
./src/styles/tailwind.css.webpack[javascript/auto]!=!./node_modules/next/dist/build/webpack/loaders/css-loader/src/index.js??ruleSet[1].rules[14].oneOf[12].use[2]!./node_modules/next/dist/build/webpack/loaders/postcss-loader/src/index.js??ruleSet[1].rules[14].oneOf[12].use[3]!./src/styles/tailwind.css
./src/styles/tailwind.css
 GET / 500 in 25367ms
 GET / 500 in 25371ms

```

If applicable, propose a fix immediately.